điều chỉnh entity
viết lại các giá trị entity cho phù hợp với database

+++ src/main/java/com/poly/demo/entity/Ticket.java
@@ -13,7 +13,7 @@ public class Ticket {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer ticketId;
 
-    private BigDecimal price;
+    private Integer price;
 
     @Enumerated(EnumType.STRING)
     private TicketStatus ticketStatus;
